requirements.txt
rich>=13.0
python-dotenv>=1.0
kokoro>=0.9.4
soundfile
pydub>=0.25
pytest
